FL2000 register access Wireshark dissector
This initial implementation only support decoding of USB_CONTROL
messages that implement FL2000 read/write operations. Unfortunately, in
order to implement this we need to register in 'usb.control' dissector
table of Wireshark which does not support dissecting control messages
for which recepient is "DEVICE". In order to overcome this a patch
needed for USB packet dissector.

TODO:
- implement I2C operations in general
- implement brifge register access for e.g. IT66121 over I2C
- implement EDID reading