Put it to compile under 5.10 kernel version
All the alterations cited here happened between 5.8 and 5.10 kernel
versions.

drm_dev_init() is not a "public" function anymore, so drm_dev_alloc() or
devm_drm_dev_alloc() must be used instead. And they use a different form
of initialization, so a lot of things have moved around or had just been
cutted off.

Why devm_drm_dev_alloc()? It was the most popular in drivers/gpu/drm
drivers.

drm_gem_object_put() is the replacement to the _unlocked() sufixed
option. The latter took care of locks, and that, as far as I understood,
is not need anymore.

drm_prime_pages_to_sg() changed it's signature. Now it needs a pointer
to the device struct as it's first argument. Luckly, the drm_gem_object
struct carries a pointer to it ;)